requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.2.0
scipy>=1.7.0
matplotlib>=3.4.2
tqdm>=4.61.1

# Web UI dependencies
Flask>=2.0.1
Flask-Cors>=3.0.10
Werkzeug>=2.0.1
requests>=2.25.1

# Image processing
opencv-python>=4.5.2

# Utilities
pyyaml>=5.4.1
click>=8.0.1
loguru>=0.5.3

# Optional: for model optimization
onnx>=1.9.0
onnxruntime>=1.8.1
